// src/interfaces/edgePipeline.types.ts
export interface EdgePipelineRecord {
    auctionName: string;
    pictureCount: number;
    runNumber: string;
    stockNumber: string;
    year: number;
    make: string;
    model: string;
    style: string;
    color: string;
    odometer: number;
    cr: string;
    grade: number;
    saleDate: string;
    lane: string;
    vin: string;
    soldAmount: string;
    watchNotes: string;
  }